requirements.txt
django
djangorestframework
psycopg2-binary  # Or psycopg2 if binary doesn't work
stripe